start css script, group by component folder
