+++ package.json
@@ -83,6 +83,7 @@
     "eslint-plugin-prettier": "4.2.1",
     "eslint-plugin-react": "7.31.6",
     "eslint-plugin-react-hooks": "4.6.0",
+    "esno": "^0.17.0",
     "fs-extra": "^11.1.1",
     "identity-obj-proxy": "3.0.0",
     "jest": "^29.0.3",
