+++ packages/css/package.json
@@ -2,7 +2,7 @@
   "name": "@digdir/designsystemet-css",
   "version": "1.0.0",
   "description": "CSS files for the Digdir Common Design System",
-  "main": "dist/index.cjs.js",
+  "main": "index.css",
   "repository": "https://github.com/digdir/designsystem",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
