+++ packages/react/package.json
@@ -11,7 +11,7 @@
     "dist/**"
   ],
   "scripts": {
-    "build": "yarn run clean && tsc -p tsconfig.build.json && yarn run copy-files && rollup -c --bundleConfigAsCjs",
+    "build": "yarn run clean && tsc -p tsconfig.build.json && esno ./scripts/generate-css.ts && yarn run copy-files && rollup -c --bundleConfigAsCjs",
     "clean": "rimraf dist && rimraf tsc-build",
     "copy-files": "copyfiles -u 1 ./src/**/*.css ./tsc-build/"
   },
